pyproject.toml
[tool.poetry]
name = "course-work-6"
version = "0.1.0"
description = ""
authors = ["Slava Lukin>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
isort = "^5.13.2"
black = "^24.8.0"
ipython = "^8.27.0"
django-filter = "^24.3"
coverage = "^7.6.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
requests = "^2.32.3"
celery = "^5.4.0"
eventlet = "^0.36.1"
redis = "^5.0.8"
django-celery-beat = "^2.7.0"
djangorestframework-simplejwt = "^5.3.1"
resources = "^0.0.1"
rest-framework-simplejwt = "^0.0.2"
setuptools = "^74.0.0"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
